+++ apps/core/storage/db/organization.sql.go
@@ -92,8 +92,8 @@ LIMIT $1 OFFSET $2
 `
 
 type ListOrganizationsParams struct {
-	Limit  int64 `json:"limit"`
-	Offset int64 `json:"offset"`
+	Limit  int32 `json:"limit"`
+	Offset int32 `json:"offset"`
 }
 
 func (q *Queries) ListOrganizations(ctx context.Context, arg ListOrganizationsParams) ([]Organization, error) {
